+++ rotkehlchen/chain/ethereum/modules/kyber/accountant.py
@@ -3,7 +3,7 @@ from typing import TYPE_CHECKING
 from rotkehlchen.accounting.structures.base import get_tx_event_type_identifier
 from rotkehlchen.accounting.structures.types import HistoryEventSubType, HistoryEventType
 from rotkehlchen.chain.evm.accounting.interfaces import ModuleAccountantInterface
-from rotkehlchen.chain.evm.accounting.structures import TxEventSettings, TxMultitakeTreatment
+from rotkehlchen.chain.evm.accounting.structures import TxAccountingTreatment, TxEventSettings
 
 from .constants import CPT_KYBER
 
@@ -21,7 +21,6 @@ class KyberAccountant(ModuleAccountantInterface):
                 count_entire_amount_spend=False,
                 count_cost_basis_pnl=True,
                 method='spend',
-                take=2,
-                multitake_treatment=TxMultitakeTreatment.SWAP,
+                accounting_treatment=TxAccountingTreatment.SWAP,
             ),
         }