+++ rotkehlchen/chain/ethereum/modules/zksync/accountant.py
@@ -21,7 +21,6 @@ class ZksyncAccountant(ModuleAccountantInterface):
                 count_entire_amount_spend=False,
                 count_cost_basis_pnl=False,
                 method='spend',
-                take=1,
-                multitake_treatment=None,
+                accounting_treatment=None,
             ),
         }