+++ rotkehlchen/chain/optimism/modules/optimism_bridge/accountant.py
@@ -20,15 +20,13 @@ class OptimismBridgeAccountant(ModuleAccountantInterface):
                 count_entire_amount_spend=False,
                 count_cost_basis_pnl=True,
                 method='spend',
-                take=1,
-                multitake_treatment=None,
+                accounting_treatment=None,
             ),
             get_tx_event_type_identifier(HistoryEventType.WITHDRAWAL, HistoryEventSubType.BRIDGE, CPT_OPTIMISM): TxEventSettings(  # noqa: E501
                 taxable=False,
                 count_entire_amount_spend=False,
                 count_cost_basis_pnl=True,
                 method='acquisition',
-                take=1,
-                multitake_treatment=None,
+                accounting_treatment=None,
             ),
         }